requirements.txt
jira
jupyter
matplotlib
pandas
tqdm
python-dateutil